+++ packages/core/lib/core/component.ts
@@ -1,4 +1,4 @@
-import type { ComponentProps, ComponentSetup, RefElement } from "../types";
+import type { ComponentSetup, RefElement } from "../types";
 
 export function defineComponent<
   SetupResult extends Record<string, unknown> | void,
@@ -6,14 +6,14 @@ export function defineComponent<
 >(opts: {
   name: string;
   props: Props;
-  setup(el: RefElement, props: ComponentProps<Props>): SetupResult;
+  setup(el: RefElement, props: Readonly<Props>): SetupResult;
 }): ComponentSetup<SetupResult, Props>;
 export function defineComponent<
   SetupResult extends Record<string, unknown> | void,
   Props extends Record<string, unknown>,
 >(opts: {
   name: string;
-  setup(el: RefElement, props: ComponentProps<Props>): SetupResult;
+  setup(el: RefElement, props: Readonly<Props>): SetupResult;
 }): ComponentSetup<SetupResult, Props>;
 export function defineComponent<
   SetupResult extends Record<string, unknown> | void,
