+++ packages/core/types/core/_internal/errorReport.d.ts
@@ -1,0 +1,5 @@
+import { type LifecycleErrorDetails } from "../error";
+import type { ComponentContextImpl } from "./component";
+type Phase = LifecycleErrorDetails["phase"];
+export declare function errorReport(phase: Phase, target: ComponentContextImpl, cause: unknown, parent?: ComponentContextImpl | null): void;
+export {};
