+++ packages/core/types/hooks/core/useSlot.d.ts
@@ -1,5 +1,5 @@
 import type { ComponentContext, ComponentSetup, ExposedSetup, RefElement } from "../../types";
 export declare function useSlot(): {
     addChild<Child extends ComponentSetup>(targetOrTargets: RefElement | RefElement[], child: Child, props?: Partial<Parameters<Child["setup"]>[1]>): ComponentContext<ExposedSetup<ReturnType<Child["setup"]>>>[];
-    removeChild(children: ComponentContext[]): Promise<void>;
+    removeChild(children: ComponentContext<Record<string, unknown>>[]): Promise<void>;
 };
